Moved test file initialization code to its own function
